Attempt to fix memory corruption.
For some reason, the strings that were being passed to the old
addEvents function via the lines array would become impressively
corrupted in certain cases. The old code worked fine with a 4-line
quick test script I mocked up, but crashed horribly on an actual real
script.

I have no idea what was actually causing the strings to get corrupted,
but sending them directly to the C and copying the memory seems to fix
the problem.

Though, knowing what was actually causing the problem would be nice.
Unfortunately, this is a royal pain in the ass to debug, so I'm not
going to spend any more time trying to figure out the truth behind the
truth as long as other problems don't pop up.