Version bump.
Adds basic version info to C library, and checks for version parity
between the moonscript and C libraries.

The versions are a semantic version represented as an unsigned integer
of the form 0xMaMiPa (key: Ma = Major, Mi = Minor, Pa = Patch). Under
the current scheme, the two files must be kept at the same version for
all releases. This kind of defeats the purpose of semantic versioning,
but the system may be improved in the future.