Fix bounds calculation.
There were a bunch of problems with the previous bounds calculation
algorithm.

It didn't check to see if the x/y offset had changed after the first
image was processed.

It didn't use absolute coordinates to check for the width and height,
so they could theoretically come out wrong.

It didn't perform comparison when setting the height.