Fix bounds detection for some lines.
It turns out that dst_x and dst_y do not always accurately represent
the top-left boundary of a line.

For example, given a drawing like the following: `{\p1}m 0 0 l 0 75 m
23 18 l 23 57 58 37.5 m 75 0 l 75 75`, dst_x and dst_y will point to
(0, 0) rather than the top-left of the actual dirty glyph which is at
(23, 18).

Unfortunately this means no assumptions can be made about where the
actual minimal bounding box is without checking the entire image.

So we're back to where we were 50 commits or so ago. Oh well.