Create CRC32 hash of rasterized line.
The bounding box alone is not enough information to determine whether
or not two lines are the same. Neither is a crc32 hash, but it's a far
sight better than just having the bounding box, and it's fast and easy.

By hashing the image color, bitmap and bounding box, a reasonably
unique hash can be created.