SubInspector: improve hash as a measure of visible uniqueness.
The ASS_Image bitmap is not a reliable indication of what is actually
displayed, as it is allocated before the results are clipped to the
display resolution. Two lines could be identical in appearance due to
their differences being clipped from the view and have different hashes
because of differences in their bitmap arrays.

This commit seeks to eliminate this problem by only hashing the chunks
of the line that are visible. It is still possible to create visually 
identical lines that will not have the same hash, but it should be much
less likely for them to occur outside of lines crafted specifically to confuse SubInspector. The new method has been chosen for speed and
low memory use rather than perfect reliability.