SubInspector: set libass cache limits.
libass loves to cache things. It just really likes to cache things so
very much that it will slowly gobble up memory pretty much forever. It
has been brought to our attention that some people don't like it when
this happens for whatever reason. I mean, what is ram for if not to be
aggressively hogged by a single library? However, against our better
judgment, we find ourselves yet again kowtowing to popular opinion and
thus are attempting to reign libass in somewhat.